+++ parselcli/processors.py
@@ -7,12 +7,15 @@ from urllib.parse import urljoin
 from typing import Tuple, Union, Dict, List
 
 from bs4 import BeautifulSoup
+from requests import Response
 
 
 class Processor:
     """Base class for parselcli processors"""
 
-    def __call__(self, values: Union[List[str], str]) -> Tuple[Union[List[str], str], Dict]:
+    def __call__(
+        self, values: Union[List[str], str], response: Response = None, default: str = ""
+    ) -> Tuple[Union[List[str], str], Dict]:
         pass
 
     def __repr__(self) -> str:
@@ -25,7 +28,9 @@ class Nth(Processor):
     def __init__(self, position: int) -> None:
         self.position = int(position)
 
-    def __call__(self, values):
+    def __call__(
+        self, values: Union[List[str], str], response: Response = None, default: str = ""
+    ) -> Tuple[Union[List[str], str], Dict]:
         return values[self.position], {}
 
     def __repr__(self) -> str:
@@ -38,7 +43,9 @@ class Join(Processor):
     def __init__(self, sep=""):
         self.sep = sep
 
-    def __call__(self, values):
+    def __call__(
+        self, values: Union[List[str], str], response: Response = None, default: str = ""
+    ) -> Tuple[Union[List[str], str], Dict]:
         return self.sep.join(values), {}
 
     def __repr__(self) -> str:
@@ -52,7 +59,9 @@ class Strip(Processor):
         self.chars = chars
         super().__init__()
 
-    def __call__(self, values):
+    def __call__(
+        self, values: Union[List[str], str], response: Response = None, default: str = ""
+    ) -> Tuple[Union[List[str], str], Dict]:
         if isinstance(values, list):
             values = [v.strip(self.chars) for v in values]
             return [v for v in values if v], {}
@@ -62,7 +71,9 @@ class Strip(Processor):
 class Collapse(Processor):
     """Collapse single element lists"""
 
-    def __call__(self, values):
+    def __call__(
+        self, values: Union[List[str], str], response: Response = None, default: str = ""
+    ) -> Tuple[Union[List[str], str], Dict]:
         if isinstance(values, list) and len(values) == 1:
             return values[0], {}
         return values or "", {}
@@ -71,7 +82,9 @@ class Collapse(Processor):
 class First(Processor):
     """Take first element if possible"""
 
-    def __call__(self, values, default=""):
+    def __call__(
+        self, values: Union[List[str], str], response: Response = None, default: str = ""
+    ) -> Tuple[Union[List[str], str], Dict]:
         if isinstance(values, list):
             return values[0], {}
         return values or default, {}
@@ -80,26 +93,31 @@ class First(Processor):
 class AbsoluteUrl(Processor):
     """Urljoin element"""
 
-    def __init__(self, base):
-        self.base = base
-
-    def __call__(self, values):
+    def __call__(
+        self, values: Union[List[str], str], response: Response = None, default: str = ""
+    ) -> Tuple[Union[List[str], str], Dict]:
+        if not response:
+            return values
         if isinstance(values, list):
-            return [urljoin(self.base, v) for v in values], {}
+            return [urljoin(response.url, v) for v in values], {}
         return urljoin(self.base, values), {}
 
 
 class Len(Processor):
     """Return length"""
 
-    def __call__(self, values):
+    def __call__(
+        self, values: Union[List[str], str], response: Response = None, default: str = ""
+    ) -> Tuple[Union[List[str], str], Dict]:
         return str(len(values)), {}
 
 
 class Repr(Processor):
     """return representation of value"""
 
-    def __call__(self, values):
+    def __call__(
+        self, values: Union[List[str], str], response: Response = None, default: str = ""
+    ) -> Tuple[Union[List[str], str], Dict]:
         return repr(values), {}
 
 
@@ -119,7 +137,9 @@ class FormatHtml(Processor):
             text = "\n".join(line[1:] for line in text.splitlines()[1:-1])
         return text
 
-    def __call__(self, values):
+    def __call__(
+        self, values: Union[List[str], str], response: Response = None, default: str = ""
+    ) -> Tuple[Union[List[str], str], Dict]:
         if isinstance(values, list):
             return [self.format(element) for element in values], {}
         return self.format(values), {}
@@ -149,7 +169,9 @@ class Regex(Processor):
             return list(search.groups())
         return value
 
-    def __call__(self, values):
+    def __call__(
+        self, values: Union[List[str], str], response: Response = None, default: str = ""
+    ) -> Tuple[Union[List[str], str], Dict]:
         if isinstance(values, list):
             return [self.check(value) for value in values], {}
         return self.check(values), {}
@@ -162,7 +184,9 @@ class Slice(Processor):
         self._value = slice_range.rstrip("]")
         self.slice = slice(*(int(val) if val is not None else val for val in self._value.split(":")))
 
-    def __call__(self, values):
+    def __call__(
+        self, values: Union[List[str], str], response: Response = None, default: str = ""
+    ) -> Tuple[Union[List[str], str], Dict]:
         return values[self.slice], {}
 
     def __repr__(self) -> str:
@@ -172,7 +196,9 @@ class Slice(Processor):
 class Sum(Processor):
     """sum all values"""
 
-    def __call__(self, values):
+    def __call__(
+        self, values: Union[List[str], str], response: Response = None, default: str = ""
+    ) -> Tuple[Union[List[str], str], Dict]:
         if not isinstance(values, list):
             return values, {}
         if all(v.isdigit() for v in values):