+++ changelog.md
@@ -1,3 +1,12 @@
+UNRELEASED [1.2.0]
+- add support for rendering target in browser emulator via `--browser` flag.  
+    This feature requires optional dependency [playwright](https://github.com/microsoft/playwright-python), can be installed via `pip install parselcli[browser]`.
+    When `--browser` flag is used, a temporary playwright controlled headless browser will be launched and attached to parsel session.
+- add support for browser load instructions via `--browser-
+
+[1.1.1]
+- fix some selectors containing dash characters (`-`) being interpreted incorrectly
+
 [1.1.0]
 - add many new processors:
     ```